Refs 12051. Define file and memory shared segments
Define two new classes:
- SharedMemSegment is the same as before, using the BOOST shared memory.
- SharedFileSegment uses BOOST memory mapped files.

They both are specializations of a template class SharedSegment, which
has a base class SharedSegmentBase.

The idea is for DataSharing objects to hold a pointer to
SharedSegmentBase that will be initialized as SharedMemSegment or
SharedFileSegment depending on the configuration.

At this point no changes have been done to the users of these classes,
as they are already using SharedMemSegment

Signed-off-by: Iker Luengo <ikerluengo@eprosima.com>